Correcao controller montaria

+++ Controllers/MontariasController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using trabalho_rodeio.Models;
 
@@ -20,21 +21,22 @@ namespace trabalho_rodeio.Controllers
         // GET: Montarias
         public async Task<IActionResult> Index()
         {
-            var montarias = await _context.Montarias.Include(m => m.Peao).Include(m => m.Touro).ToListAsync();
-            return View(montarias);
+            var contexto = _context.Montarias.Include(m => m.Peao).Include(m => m.Touro);
+            return View(await contexto.ToListAsync());
         }
 
         // GET: Montarias/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Montarias == null)
             {
                 return NotFound();
             }
 
-            var montaria = await _context.Montarias.Include(m => m.Peao).Include(m => m.Touro)
+            var montaria = await _context.Montarias
+                .Include(m => m.Peao)
+                .Include(m => m.Touro)
                 .FirstOrDefaultAsync(m => m.Id == id);
-
             if (montaria == null)
             {
                 return NotFound();
@@ -46,10 +48,14 @@ namespace trabalho_rodeio.Controllers
         // GET: Montarias/Create
         public IActionResult Create()
         {
+            ViewData["PeaoId"] = new SelectList(_context.Peoes, "Id", "Nome");
+            ViewData["TouroId"] = new SelectList(_context.Touros, "Id", "Nome");
             return View();
         }
 
         // POST: Montarias/Create
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Data,PeaoId,TouroId")] Montaria montaria)
@@ -60,13 +66,15 @@ namespace trabalho_rodeio.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
+            ViewData["PeaoId"] = new SelectList(_context.Peoes, "Id", "Nome", montaria.PeaoId);
+            ViewData["TouroId"] = new SelectList(_context.Touros, "Id", "Nome", montaria.TouroId);
             return View(montaria);
         }
 
         // GET: Montarias/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Montarias == null)
             {
                 return NotFound();
             }
@@ -76,10 +84,14 @@ namespace trabalho_rodeio.Controllers
             {
                 return NotFound();
             }
+            ViewData["PeaoId"] = new SelectList(_context.Peoes, "Id", "Nome", montaria.PeaoId);
+            ViewData["TouroId"] = new SelectList(_context.Touros, "Id", "Nome", montaria.TouroId);
             return View(montaria);
         }
 
         // POST: Montarias/Edit/5
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Data,PeaoId,TouroId")] Montaria montaria)
@@ -109,18 +121,22 @@ namespace trabalho_rodeio.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
+            ViewData["PeaoId"] = new SelectList(_context.Peoes, "Id", "Nome", montaria.PeaoId);
+            ViewData["TouroId"] = new SelectList(_context.Touros, "Id", "Nome", montaria.TouroId);
             return View(montaria);
         }
 
         // GET: Montarias/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Montarias == null)
             {
                 return NotFound();
             }
 
             var montaria = await _context.Montarias
+                .Include(m => m.Peao)
+                .Include(m => m.Touro)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (montaria == null)
             {
@@ -135,19 +151,23 @@ namespace trabalho_rodeio.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_context.Montarias == null)
+            {
+                return Problem("Entity set 'Contexto.Montarias'  is null.");
+            }
             var montaria = await _context.Montarias.FindAsync(id);
             if (montaria != null)
             {
                 _context.Montarias.Remove(montaria);
-                await _context.SaveChangesAsync();
             }
-
+            
+            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool MontariaExists(int id)
         {
-            return _context.Montarias.Any(e => e.Id == id);
+          return _context.Montarias.Any(e => e.Id == id);
         }
     }
 }
